create walb device

+++ provision.sh
@@ -2,22 +2,28 @@
 
 sudo sed -i -e 's://archive://jp.archive:g' /etc/apt/sources.list
 sudo apt-get update
-sudo apt-get dist-upgrade -y
+sudo apt-get upgrade -y
 sudo apt-get install -y build-essential
 
-git clone https://github.com/hk220/osc-kyoto-walb.git
 
-WORK_DIR=$(pwd)/osc-kyoto-walb
-cd ${WORK_DIR}/walb-driver
-sudo make setup
-make
-sudo make install
-sudo make autoload
-sudo make load-module
+if [ ! -d $(pwd)/osc-kyoto-walb ]; then
+  git clone https://github.com/hk220/osc-kyoto-walb.git
 
-cd ${WORK_DIR}/walb-tools
-sudo make setup
-make
-sudo make install
+  WORK_DIR=$(pwd)/osc-kyoto-walb
+  cd ${WORK_DIR}/walb-driver
+  sudo make setup
+  make
+  sudo make install
+  sudo make autoload
+  sudo make load-module
 
-sudo reboot
+  cd ${WORK_DIR}/walb-tools
+  sudo make setup
+  make
+  sudo make install
+
+  cd ${WORK_DIR}
+  sudo sh setup_disk.sh
+  
+  sudo reboot
+fi
